extract option inteface
