--- src/Microsoft.DocAsCode.Dfm/DfmFencesBlockPathQueryRegion.cs
@@ -1,11 +1,0 @@
-﻿// Copyright (c) Microsoft. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-namespace Microsoft.DocAsCode.Dfm
-{
-    public class DfmFencesBlockPathQueryRegion
-    {
-        public int? StartLine { get; set; }
-        public int? EndLine { get; set; }
-    }
-}
