+++ src/Microsoft.DocAsCode.Dfm/Microsoft.DocAsCode.Dfm.csproj
@@ -51,7 +51,11 @@
     <Compile Include="CodeSnippetExtractor\FlatNameCodeSnippetExtractor.cs" />
     <Compile Include="CodeSnippetExtractor\RecursiveNameCodeSnippetExtractor.cs" />
     <Compile Include="DfmCodeExtractorHelper.cs" />
-    <Compile Include="DfmFencesBlockPathQueryRegion.cs" />
+    <Compile Include="DfmFencesBlockPathQueryOptions\DfmFencesBlockPathQueryOption.cs" />
+    <Compile Include="DfmFencesBlockPathQueryOptions\MultipleLineRangeBlockPathQueryOption.cs" />
+    <Compile Include="DfmFencesBlockPathQueryOptions\TagNameBlockPathQueryOption.cs" />
+    <Compile Include="DfmFencesBlockPathQueryOptions\LineRangeBlockPathQueryOption.cs" />
+    <Compile Include="DfmFencesBlockPathQueryOptions\IDfmFencesBlockPathQueryOption.cs" />
     <Compile Include="DfmMarkdownRenderer.cs" />
     <Compile Include="CodeSnippetExtractor\ICodeSnippetExtractor.cs" />
     <Compile Include="MarkdownValidators\MarkdownSytleConfig.cs" />
@@ -84,7 +88,6 @@
     <Compile Include="DfmEngine.cs" />
     <Compile Include="DfmEngineBuilder.cs" />
     <Compile Include="DfmExtractCodeResult.cs" />
-    <Compile Include="DfmFencesBlockPathQueryOption.cs" />
     <Compile Include="DfmRenderer.cs" />
     <Compile Include="DfmRendererHelper.cs" />
     <Compile Include="DocfxFlavoredIncHelper.cs" />
